Moved meta answers to Classifications instead of Annotations
