+++ src/ducks/annotations.js
@@ -11,7 +11,6 @@ const SELECT_ANNOTATION = 'SELECT_ANNOTATION';
 const UNSELECT_ANNOTATION = 'UNSELECT_ANNOTATION';
 const DELETE_SELECTED_ANNOTATION = 'DELETE_SELECTED_ANNOTATION';
 const COLLABORATE_WITH_ANNOTATION = 'COLLABORATE_WITH_ANNOTATION';
-const SET_SUBJECT_COMPLETION_ANSWERS = 'SET_SUBJECT_COMPLETION_ANSWERS';
 const UPDATE_TEXT = 'UPDATE_TEXT';
 const SAVE_TEXT = 'SAVE_TEXT';
 
@@ -39,10 +38,9 @@ const initialState = {
   annotations: [],  //Completed annotations.
   selectedAnnotation: null,  //Existing annotation that's been selected, by clicking on them. null if nothing is selected.
   selectedAnnotationIndex: null,
-  subjectCompletionAnswers: null,
 };
 
-const subjectReducer = (state = initialState, action) => {
+const annotationsReducer = (state = initialState, action) => {
   switch (action.type) {
     case RESET_ANNOTATIONS:
       return initialState;
@@ -126,11 +124,6 @@ const subjectReducer = (state = initialState, action) => {
         annotationPanePosition: null,
         selectedAnnotation: null,
         selectedAnnotationIndex: null,
-      });
-    
-    case SET_SUBJECT_COMPLETION_ANSWERS:
-      return Object.assign({}, state, {
-        subjectCompletionAnswers: action.answers,
       });
 
     default:
@@ -216,16 +209,7 @@ const collaborateWithAnnotation = (annotation, text) => {
   };
 };
 
-const setSubjectCompletionAnswers = (answers) => {
-  return (dispatch) => {
-    dispatch({
-      type: SET_SUBJECT_COMPLETION_ANSWERS,
-      answers,
-    });
-  };
-};
-
-export default subjectReducer;
+export default annotationsReducer;
 
 //------------------------------------------------------------------------------
